pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "haiip"
version = "0.1.0"
description = "Human-Aligned Industrial Intelligence Platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",

    # Database
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.29.0",

    # Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",

    # ML Core
    "scikit-learn>=1.4.2",
    "numpy>=1.26.4",
    "pandas>=2.2.2",
    "scipy>=1.13.0",
    "joblib>=1.4.2",

    # Deep Learning
    "torch>=2.3.0",
    "lightning>=2.2.0",
    "onnxruntime>=1.18.0",
    "onnx>=1.16.0",

    # RAG / LLM
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "openai>=1.30.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",

    # Data Ingestion
    "asyncua>=1.1.0",
    "aiomqtt>=2.0.0",

    # Observability
    "prometheus-fastapi-instrumentator>=7.0.0",
    "structlog>=24.2.0",

    # Task Queue
    "celery>=5.4.0",
    "redis>=5.0.4",

    # Dataset Loaders
    "ucimlrepo>=0.0.6",
    "requests>=2.32.0",

    # Dashboard
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
    "altair>=5.3.0",

    # Explainability
    "shap>=0.45.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.5.0",
    "ruff>=0.4.7",
    "mypy>=1.10.0",
    "pre-commit>=3.7.1",
    "bandit>=1.7.8",
]

[tool.hatch.build.targets.wheel]
packages = ["haiip"]

# ── Ruff (linter + formatter) ─────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B", "C4", "UP"]
ignore = [
    "S101",  # assert — used in tests and internal guards
    "S105",  # hardcoded-password-string — token labels are not secrets
    "S110",  # try-except-pass — intentional for optional-dependency imports
    "S311",  # pseudo-random — not used for cryptographic purposes
    "S310",  # suspicious-urllib-open — internal HTTP only
    "B905",  # zip-without-explicit-strict — fine on Python 3.11+
    "E501",  # line-too-long — formatter handles; long lines in ML/generated code
]

[tool.ruff.lint.per-file-ignores]
"haiip/tests/**" = ["S101", "S106", "S108", "S603", "S607"]

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["haiip/tests"]
addopts = "--cov=haiip --cov-report=term-missing --cov-fail-under=80"

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["haiip"]
omit = ["tests/*", "notebooks/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
